fixed constants

+++ Runtime/Scripts/Constants/FloatingPoints/Float.cs
@@ -11,11 +11,11 @@
 		public const float Zero = Byte.Zero;
 		public const float One = Byte.One;
 
-		public const float OneHalf = (float)Decimal.OneHalf;
+		public const float OneHalf = (float)Double.OneHalf;
 
-		public const float FullCircleDegrees = (float)Decimal.FullCircleDegrees;
-		public const float HalfCircleDegrees = (float)Decimal.HalfCircleDegrees;
-		public const float Rad2Deg = (float)Decimal.Rad2Deg;
-		public const float Deg2Rad = (float)Decimal.Deg2Rad;
+		public const float FullCircleDegrees = (float)Double.FullCircleDegrees;
+		public const float HalfCircleDegrees = (float)Double.HalfCircleDegrees;
+		public const float Rad2Deg = (float)Double.Rad2Deg;
+		public const float Deg2Rad = (float)Double.Deg2Rad;
 	}
 }
